Move auth functions to auth.ts
Including:
REDIRECT_URI_OOB
oauth2Client
getAPICredentials
authorizeWithLocalHost
authorizeWithoutLocalHost
